added admin restrictions to CRUD, added admin roles to user tables, updated navbar to check user logged in status, added a register option on the navbar

+++ app/Http/Controllers/BeachController.php
@@ -14,6 +14,9 @@ class BeachController extends Controller
 //   show all beaches
     public function index()
     {
+
+        // no admin check as this page is for all users 
+        
         // beach fetcher
        $beaches = Beach::all();
         // show the beaches
@@ -23,12 +26,21 @@ class BeachController extends Controller
     // show the create form
     public function create()
     {
+        // check if the user is an admin
+        if (auth()->user()->role !== 'admin') {
+            abort(403);
+        }
         return view('beaches.create');
     }
 
     // save a new beach to the database
     public function store(Request $request)
     {
+        // check if the user is an admin
+        if (auth()->user()->role !== 'admin') {
+            abort(403);
+        }
+
         $request->validate([
             'name' => 'required',
             'description' => 'required',
@@ -45,18 +57,29 @@ class BeachController extends Controller
     // show each beach
     public function show(Beach $beach)
     {
+        // no admin check as all users are allowed to view beaches
         return view ('beaches.show', compact('beach'));
     }
 
     //show the edit form
     public function edit(Beach $beach)
     {
+        // check if the user is an admin
+        if (auth()->user()->role !== 'admin') {
+            abort(403);
+        }
+
         return view ('beaches.edit', compact('beach'));
     }
 
     // update the beaches
     public function update(Request $request, Beach $beach)
     {
+        // check if the user is an admin
+        if (auth()->user()->role !== 'admin') {
+            abort(403);
+        }
+
         $request->validate([
             'name' => 'required',
             'description' => 'required',
@@ -74,6 +97,11 @@ class BeachController extends Controller
     // delete any beach
     public function destroy(Beach $beach)
     {
+        // check if the user is an admin
+        if (auth()->user()->role !== 'admin') {
+            abort(403);
+        }
+
         $beach->delete();
 
         return redirect()->route('beaches.index')->with('success', 'Beach deleted successfully!');
